+++ src/allotropy/parsers/agilent_gen5_image/constants.py
@@ -7,7 +7,6 @@ from allotropy.allotrope.models.adm.plate_reader.benchling._2023._09.plate_reade
 MULTIPLATE_FILE_ERROR = "Only a single plate per file can be processed at this time. Please refer to Gen5 documentation for how to generate single plate exports from multi-plate experiments"
 NO_PLATE_DATA_ERROR = "No plate data found in file."
 UNSUPPORTED_READ_TYPE_ERROR = "Only imaging results can be processed at this time."
-DEFAULT_EXPORT_FORMAT_ERROR = "Could not find 'Results' section. This export format cannot be processed at this time - ensure the 'Regroup data in one matrix/table' option is enabled within Gen5"
 
 HEADER_PREFIXES = frozenset(
     {
@@ -26,6 +25,8 @@ HEADER_PREFIXES = frozenset(
 DEFAULT_SOFTWARE_NAME = "Gen5 Image"
 DEVICE_TYPE = "Imager"
 DETECTION_TYPE = "optical-imaging"
+METADATA_ONLY = "METADATA ONLY"
+POSSIBLE_WELL_COUNTS = [1, 2, 4, 6, 8, 12, 24, 48, 72, 96, 384, 1536, 3456]
 
 AUTOFOCUS_STRINGS = frozenset(
     {